more config for i3 and add feature to song_manager

+++ song_manager/main.py
@@ -1,21 +1,17 @@
+from itertools import filterfalse
 from bs4 import BeautifulSoup
 from mutagen.easyid3 import EasyID3
 from csv import reader
 from urllib.parse import quote
 import youtube_dl
-import os, sys
-import time
-import argparse
+import os
 import requests
 import shutil
 
-# TODO error handle 
 
 class Downloader():
-    def __init__(self, sdir, log_file, force_download):
+    def __init__(self, sdir):
         self.sdir = sdir
-        self.log_file = log_file
-        self.force_download = force_download
         self.slink = ''
         self.dlink = ''
 
@@ -32,34 +28,27 @@ class Downloader():
         audio['artist'] = self.name[1]
         audio.save()
 
-    def write_log(self, error=''):
-        timestamp = time.asctime()
-        template = f"[{timestamp}] {self.song_name}\n[Searching link]: {self.slink}\n[Downloading link]: {self.dlink}\n{error}\n"
-        self.log_file.write(template)
-
 
 class DefaultDownloader(Downloader):
-    def __init__(self, sdir, log_file, force_download=False):
-        super().__init__(sdir, log_file, force_download)
+    def __init__(self, sdir):
+        super().__init__(sdir)
 
     def manage(self, song_name, pos, search_by_title_only=False):
         super().process_song_name(song_name, pos)
 
-        if self.force_download or not os.path.isfile(self.song_path):
+        if not os.path.isfile(self.song_path):
             self.get_song_link(search_by_title_only)
             self.get_download_link()
             self.download()
             self.change_metadata()
-            self.write_log()
 
     def manage_link(self, song_name, link):
         super().process_song_name(song_name)
         self.slink = link
-        if self.force_download or not os.path.isfile(self.song_path):
+        if not os.path.isfile(self.song_path):
             self.get_download_link()
             self.download()
             self.change_metadata()
-            self.write_log()
 
     def get_song_link(self, search_by_title_only):
         if search_by_title_only is True:
@@ -97,26 +86,24 @@ class YoutubeDownloader(Downloader):
         }],
     }
 
-    def __init__(self, sdir, log_file, force_download=False):
-        super().__init__(sdir, log_file, force_download)
+    def __init__(self, sdir):
+        super().__init__(sdir)
 
     def manage(self, song_name, pos=1):
         super().process_song_name(song_name, pos)
 
-        if self.force_download or not os.path.isfile(self.song_path):
+        if not os.path.isfile(self.song_path):
             self.get_youtube_link()
             self.download()
             self.change_metadata()
-            self.write_log()
     
     def manage_link(self, song_name, link):
         super().process_song_name(song_name)
         self.dlink = link
         self.slink = ''
-        if self.force_download or not os.path.isfile(self.song_path):
+        if not os.path.isfile(self.song_path):
             self.download()
             self.change_metadata()
-            self.write_log()
 
     def get_youtube_link(self):
         searching_link = "https://youtube.com/results?search_query={}".format(quote(' '.join(self.name)))
@@ -146,10 +133,6 @@ class YoutubeDownloader(Downloader):
 def download_song(song_name, tag, pos):
     if not tag:
         default_downloader.manage(song_name, pos)
-    elif tag == 'dn':
-        default_downloader.manage(song_name, pos, True)
-    elif tag == 'y':
-        youtube_downloader.manage(song_name, pos)
     elif tag == 'dl':
         default_downloader.manage_link(song_name, pos)
     elif tag == 'yl':
@@ -157,40 +140,47 @@ def download_song(song_name, tag, pos):
 
 def delete_song(song_dir, song_list):
     for file_name in os.listdir(song_dir):
-        if file_name.endswith(".mp3"):
+        if file_name.endswith(".mp3") and file_name not in song_list:
             file_name_path = os.path.join(song_dir, file_name)
-            if file_name not in song_list:
-                os.remove(file_name_path)
-                print('\033[1;36m' + "Delete {}".format(file_name) + '\033[1;36m')
+            os.remove(file_name_path)
+            print('\033[1;36m' + "Delete {}".format(file_name) + '\033[1;36m')
 
 def copy(from_dir, end_dir):
     end_dir_file_name = os.listdir(end_dir)
     for file_name in os.listdir(from_dir):
-        if file_name.endswith(".mp3"):
+        if file_name.endswith(".mp3") and file_name not in end_dir_file_name:
             file_name_path = os.path.join(from_dir, file_name)
-            if file_name not in end_dir_file_name:
-                shutil.copyfile(file_name_path, os.path.join(end_dir, file_name))
-                print('\033[1;33m' + "Copy {}".format(file_name) + '\033[1;33m')
+            shutil.copyfile(file_name_path, os.path.join(end_dir, file_name))
+            print('\033[1;33m' + "Copy {}".format(file_name) + '\033[1;33m')
+
+def add_to_old_list(old_list_dir, song_list_dir):
+    with open(song_list_dir) as song_list_name_something, open(old_list_dir, "r+") as old_list_name_something:
+        songs = reader(song_list_name_something, delimiter=',')
+        for s in songs:
+            if (s and not s[0].startswith("#")):
+                old_list_name_something.seek(0)
+                old_list = reader(old_list_name_something, delimiter=",")
+                found = False
+                for ol in old_list:
+                    if (s[0] == ol[0]):
+                        found = True
+                        break
+
+                if (not found):
+                    old_list_name_something.write(','.join(s)+'\n')
 
 if __name__ == "__main__":
-    parser = argparse.ArgumentParser(description="A small script to download songs")
-    parser.add_argument("--song-dir", default=os.path.join(os.environ['HOME'], 'Music'), help="provide download directory (default: ~/Music")
-    parser.add_argument("-r", action="store_true")
-    my_parse = parser.parse_args()
-
-    if os.path.exists("log"):
-        log_file = open("log", 'a')
-    else:
-        log_file = open("log", 'w')
-
     os.system("sudo umount /mnt")
     os.system("sudo jmtpfs -o allow_other /mnt")
     print('\033[1;36m' + "DONE" + '\033[1;36m')
 
     song_list = []
-    default_downloader = DefaultDownloader(my_parse.song_dir, log_file, my_parse.r)
-    youtube_downloader = YoutubeDownloader(my_parse.song_dir, log_file, my_parse.r)
+    song_dir = os.path.join(os.environ["HOME"], "Music")
+
+    default_downloader = DefaultDownloader(song_dir)
+    youtube_downloader = YoutubeDownloader(song_dir)
     dir_path = os.path.dirname(os.path.realpath(__file__))
+
     with open(os.path.join(dir_path, "song_list")) as file:
         csv_reader = reader(file, delimiter=',')
         for s in csv_reader:
@@ -202,32 +192,34 @@ if __name__ == "__main__":
                     pos = int(s[2])
 
                 download_song(s[0], s[1], pos)
-    delete_song(my_parse.song_dir, song_list)
+    delete_song(song_dir, song_list)
 
     path = r"/mnt/Internal shared storage/Music/Music"
     delete_song(path, song_list)
     # Copy to phone
-    copy(my_parse.song_dir, path)
+    copy(song_dir, path)
 
-    song_list = []
-    new_dir = os.path.join(os.environ["HOME"], "PhoneMusic")
-    default_downloader.sdir = new_dir
-    youtube_downloader.sdir = new_dir 
-    with open(os.path.join(dir_path, "phone_list")) as file:
-        csv_reader = reader(file, delimiter=',')
-        for s in csv_reader:
-            if s and (not s[0].startswith('#')):
-                pos = 1
-                if s[1] == 'dl' or s[1] == 'yl':
-                    pos = s[2]
-                elif s[2]:
-                    pos = int(s[2])
+    add_to_old_list(os.path.join(dir_path, "old_list"), os.path.join(dir_path, "song_list"))
 
-                download_song(s[0], s[1], pos)
-    delete_song(new_dir, song_list)
+    # song_list = []
+    # new_dir = os.path.join(os.environ["HOME"], "PhoneMusic")
+    # default_downloader.sdir = new_dir
+    # youtube_downloader.sdir = new_dir 
+    # with open(os.path.join(dir_path, "phone_list")) as file:
+    #     csv_reader = reader(file, delimiter=',')
+    #     for s in csv_reader:
+    #         if s and (not s[0].startswith('#')):
+    #             pos = 1
+    #             if s[1] == 'dl' or s[1] == 'yl':
+    #                 pos = s[2]
+    #             elif s[2]:
+    #                 pos = int(s[2])
 
-    path = r"/mnt/Internal shared storage/Music/Long one"
-    delete_song(path, song_list)
-    # copy to phone
-    copy(new_dir, path)
+    #             download_song(s[0], s[1], pos)
+    # delete_song(new_dir, song_list)
+
+    # path = r"/mnt/Internal shared storage/Music/Long one"
+    # delete_song(path, song_list)
+    # # copy to phone
+    # copy(new_dir, path)
